6601904: vm complain about "foundation.jar" is not a valid classpath zip file or directory"
The fix is to not add foundation.jar to CVM_JARFILES when CVM_PRELOAD_LIB=true. The jar file does not exist in this case, because the classes are part of btclasses.zip instead. Thus we don't want foundation.jar on the bootclasspath.

Reviewed by Hinkmond and Mark
Did test builds and TCK on wm5.