Fix the following assert while running the tck with a nullapp build:
Assertion failed at line 1813 in ../../src/share/javavm/runtime/split_verify.c: CVMcpTypeIs(constPool, memberClassIndex, ClassTypeID)
 
The following line in vIsProtectedAccess caused the assert:

    CVMClassTypeID tClass = CVMcpGetClassTypeID(constPool, memberClassIndex);

The problem is that the split verifier is assuming that no ClassRef entry has been fully resolved, so it must contains a CVMClassTypeID. However, cvm does some eager cp resolution just after class loading, and in this case the ClassRef had been fully resolved to a cb. The fix is to check if the cp entries is resolved already, and get the typeid from the cb if it is.

Reviewed by Bill
Tested with CDC TCK